+++ test/unit/export/console_exporter_test.dart
@@ -9,7 +9,6 @@
 import 'dart:async';
 
 import 'package:dartastic_opentelemetry/dartastic_opentelemetry.dart';
-import 'package:dartastic_opentelemetry_api/dartastic_opentelemetry_api.dart';
 import 'package:test/test.dart';
 
 /// Captures console output for testing