+++ test/unit/logs/logs_coverage_test.dart
@@ -16,7 +16,6 @@
 import 'dart:async';
 
 import 'package:dartastic_opentelemetry/dartastic_opentelemetry.dart';
-import 'package:dartastic_opentelemetry_api/dartastic_opentelemetry_api.dart';
 import 'package:fixnum/fixnum.dart';
 import 'package:test/test.dart';