+++ test/unit/otel_sdk_test.dart
@@ -11,7 +11,6 @@
 import 'dart:typed_data';
 
 import 'package:dartastic_opentelemetry/dartastic_opentelemetry.dart';
-import 'package:dartastic_opentelemetry_api/dartastic_opentelemetry_api.dart';
 import 'package:test/test.dart';
 
 import '../testing_utils/in_memory_span_exporter.dart';