+++ test/unit/sdk/resource_detector_test.dart
@@ -4,7 +4,6 @@
 import 'dart:io' as io;
 
 import 'package:dartastic_opentelemetry/dartastic_opentelemetry.dart';
-import 'package:dartastic_opentelemetry_api/dartastic_opentelemetry_api.dart';
 import 'package:test/test.dart';
 
 /// A custom detector that returns a resource with the given attributes.