+++ test/unit/trace/sampling/trace_id_ratio_sampler_test.dart
@@ -4,7 +4,6 @@
 import 'dart:math';
 
 import 'package:dartastic_opentelemetry/dartastic_opentelemetry.dart';
-import 'package:dartastic_opentelemetry_api/dartastic_opentelemetry_api.dart';
 import 'package:test/test.dart';
 
 void main() {