+++ test/unit/util/otel_env_test.dart
@@ -3,8 +3,6 @@
 
 import 'package:dartastic_opentelemetry/dartastic_opentelemetry.dart';
 import 'package:dartastic_opentelemetry/src/trace/span_logger.dart';
-import 'package:dartastic_opentelemetry_api/dartastic_opentelemetry_api.dart'
-    show LogFunction, LogLevel;
 import 'package:test/test.dart';
 
 void main() {